+++ R/dittoViz_yPlot_module_server.R
@@ -486,10 +486,16 @@ dittoViz_yPlotServer <- function(id, data, hide.inputs = NULL, hide.tabs = NULL,
             return(x)
         })
 
-        output$download.interactive.summary <- create_interactive_summary_download_handler(
-            plot_reactive = generate_yPlot,
-            stats_reactive = last_stats_df,
-            inputs_reactive = AllInputs(),
+        plot_summary_reactive <- reactive({
+            create_interactive_summary_data(
+                plot_reactive = generate_yPlot,
+                stats_reactive = last_stats_df,
+                inputs_reactive = AllInputs()
+            )
+        })
+
+        output$download.interactive.summary <- .create_download_file(
+            data_list = plot_summary_reactive,
             filename_base = "yPlot_summary"
         )
 
@@ -506,5 +512,7 @@ dittoViz_yPlotServer <- function(id, data, hide.inputs = NULL, hide.tabs = NULL,
                 )
             }
         )
+
+        return(plot_summary_reactive)
     })
 }