+++ R/dumbbellPlot_module_server.R
@@ -301,10 +301,18 @@ dumbbellPlotServer <- function(id, data, hide.inputs = NULL, hide.tabs = NULL, d
             return(x)
         })
 
-        output$download.interactive.summary <- create_interactive_summary_download_handler(
-            plot_reactive = generate_dumbbellPlot,
-            inputs_reactive = AllInputs(),
+        plot_summary_reactive <- reactive({
+            create_interactive_summary_data(
+                plot_reactive = generate_dumbbellPlot,
+                inputs_reactive = AllInputs()
+            )
+        })
+
+        output$download.interactive.summary <- .create_download_file(
+            data_list = plot_summary_reactive,
             filename_base = "dumbbellPlot_summary"
         )
+
+        return(plot_summary_reactive)
     })
 }