+++ R/plotthis_AreaPlot_module_server.R
@@ -268,10 +268,18 @@ plotthis_AreaPlotServer <- function(id, data, hide.inputs = NULL, hide.tabs = NU
             return(x)
         })
 
-        output$download.interactive.summary <- create_interactive_summary_download_handler(
-            plot_reactive = generate_AreaPlot,
-            inputs_reactive = AllInputs(),
+        plot_summary_reactive <- reactive({
+            create_interactive_summary_data(
+                plot_reactive = generate_AreaPlot,
+                inputs_reactive = AllInputs()
+            )
+        })
+
+        output$download.interactive.summary <- .create_download_file(
+            data_list = plot_summary_reactive,
             filename_base = "AreaPlot_summary"
         )
+
+        return(plot_summary_reactive)
     })
 }